feat: Add GPU Offer Search and Instance Management Modals
- Implemented GpuOfferSearchModal for searching and selecting GPU offers with VLLM configuration.
- Added InstanceManagementModal for managing instances, including loading, deleting, and configuring VLLM settings.
- Integrated API calls for searching offers and managing instances.
- Enhanced UI with loading states, error handling, and user feedback through toasts.
